+++ plugins/search/providers/sphinx_search/lib/kSphinxSearchManager.php
@@ -213,7 +213,7 @@ class kSphinxSearchManager implements kObjectUpdatedEventConsumer, kObjectAddedE
 			$getter = "get" . $getterName;
 			$fieldType =  $objectIndexClass::getFieldType($field);
 			$nullable = $objectIndexClass::isNullableField($field);
-			
+
 			switch($fieldType)
 			{
 				case IIndexable::FIELD_TYPE_STRING:
@@ -221,13 +221,25 @@ class kSphinxSearchManager implements kObjectUpdatedEventConsumer, kObjectAddedE
 					if($nullable)
 					{
 						if(is_null($value) || $value === '')
-							$value = self::HAS_NO_VALUE . $object->getPartnerId();
+						{
+							$suffix = ' '.self::HAS_NO_VALUE.$object->getPartnerId();
+						}
 						else
-							$value .= ' ' . self::HAS_VALUE . $object->getPartnerId();
+						{
+							$suffix = ' '.self::HAS_VALUE.$object->getPartnerId();
+						}
+
+						$enricher = "enrich" . $getterName;
+						if(is_callable(array($object, $enricher)))
+						{
+							$value = $object->$enricher($value);
+						}
+
+						$value.=$suffix;
 					}
+
 					$dataStrings[$field] = $value;
 					break;
-					
 				case IIndexable::FIELD_TYPE_INTEGER:
 					$dataInts[$field] = $object->$getter();
 					break;